refactoring directory structure -- renaming files
