+++ controllers/VocabulariesController.go
